[FEATURE] Inject workstation's chef credentials into test-kitchen VM
As we are testing our CI pipeline, we need access to the Chef server at
some points (berks resolution, cookbook upload).
To avoid creation of test/dummy users, we can inject the credentials of
the dude running `kitchen converge` into the VM so that it has full horse
power.

This can be controlled in the .kitchen.yml, optionally extended / disabled
by default.