[BUGFIX] Don't write job template file to file_cache_path
As described in someara/kitchen-dokken#118, the contents of
Chef::Config[:file_cache_path] are persistent on the host's file system.

This leads to the unexpected effect that the Jenkins job not reliably created,
if the previous run of kitchen-dokken had left over an identical template file.

Therefore, put the file into a directory inside the container, which is
reliably destroyed together with the container.